test(logger): splitting existing tests to avoid pipeline failing for timeout
